# doc/swagger.json
{"swagger":"2.0","info":{"title":"markforged","version":"1.0.0","description":"Markforged API exercise"},"paths":{"/objects":{"post":{"tags":["Objects"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/AddObject"}}]},"delete":{"tags":["Objects"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"name","in":"path","required":true,"type":"string","description":"The unique name of the object to be deleted "}]},"get":{"tags":["Objects"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"name","in":"path","required":true,"type":"string","description":"The unique name of the object "}]},"put":{"tags":["Objects"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UpdateObject"}}]}},"/prints":{"post":{"tags":["_Users_Hugh_Documents_work_markforged_src_api_controller_printerController_js"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/AddPrint"}}]}},"/prints/status":{"get":{"tags":["_Users_Hugh_Documents_work_markforged_src_api_controller_printerController_js"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"status","in":"path","required":true,"type":"string","description":"The status "}]}}},"definitions":{"AddObject":{"properties":{"name":{"type":"string","description":"the unique name of the object "},"materialType":{"type":"string","description":"the materialType of the object "},"printerType":{"type":"string","description":"the printerType of the object "},"orientation":{"type":"string","description":"the orientation of the object "},"density":{"type":"string","description":"the density of the object "}},"required":["name","materialType","printerType","orientation","density"]},"DeleteObject":{"properties":{"name":{"type":"string","description":"The unique name of the object to be deleted "}},"required":["name"]},"GetObjectsByName":{"properties":{"name":{"type":"string","description":"The unique name of the object "}},"required":["name"]},"UpdateObject":{"properties":{"name":{"type":"string","description":"The unique name of the object to be updated "},"materialType":{"type":"string","description":"optional: The new materialType of the objet "},"printerType":{"type":"string","description":"optional: The new printerType of the object "},"orientation":{"type":"string","description":"optional: The new orientation of the object "},"density":{"type":"string","description":"optional: The new density of the object "}},"required":["name","materialType","printerType","orientation","density"]},"AddPrint":{"properties":{"objectName":{"type":"string","description":"the name of the object to print "},"printer":{"type":"string","description":"the name of the printer to print with "}},"required":["objectName","printer"]},"GetPrintsByStatus":{"properties":{"status":{"type":"string","description":"The status "}},"required":["status"]}}}